Service method for creating new project
After creation of new project (and storing it in database) read model
database also needs insert of new project.

Service invokes all SQL queries and updates in a transaction. It also
checks if there is no project with given `projectId` already in database
(value of `projectId` needs to be unique).