Update statistics after adding task
Statistics entity changed to be able to update it after task remove.
Instead of average we keep in database sums and number of task,
(which gives us possibility to compute average).

Task can be started at any time and have arbitrary duration, statistics
are updated for each month that this task time span overlaps.